Inline CountingConcourse into CountingConcourseConnectionPool
